Users/mbarbour/update teams extension for missingparts (#246)
* Fix for null check.. adding in global routes for teams.

* Add TeamsEnableFeedbackLoop method for feedback handling

Introduces the `TeamsEnableFeedbackLoop` method in the
`Microsoft.Agents.Extensions.Teams` namespace. This method
allows adding a feedback loop flag to the `ChannelData`
object of an `IActivity`. It includes XML documentation
and checks if `ChannelData` is already populated before
creating a new object. Returns `false` if the data is
already present.