Updated Teams AI Lib to comply with project specs and create a proper… (#306)
* Updated Teams AI Lib to comply with project specs and create a proper nuget packge.

* Update NoWarn property to include CA1854

Added CA1854 to the NoWarn property in the project file
to suppress warnings related to this specific code analysis
rule, in addition to the existing warnings.

* Refactor code to enhance functionality and remove redundancies

This commit introduces new features and improvements by adding essential lines of code while removing outdated or unnecessary sections. The changes aim to optimize the program's behavior and maintainability.

* Remove readme.md from packaging in Teams AI project

* Add README.md to packaging for Teams AI library